Fix 'games' data in localStorage
Since commit 5fe0d7af1abc6b6b66eed1a51186902b5abed0c0, the 'games'
localStorage value was initialized as an array when it needed to be an
object. How this issue persisted for so long without any notice while
work was still being done is beyond me. But the issue is fixed and it
works as intended.